Finished feature join by link
